fix: formatting and clippy assign_op_pattern in benchmarks
Fix CI failures from ContextMessage type change in context_benchmark.rs:
- Unwrap ContextMessage::Event before passing to process/dispatch_await
- Fix unnecessary_cast warnings in checkpoint and watermark tests
- Fix for_kv_map warning in join.rs checkpoint restore

+++ crates/varpulis-runtime/tests/watermark_tests.rs
@@ -90,7 +90,7 @@ async fn test_watermark_advance_triggers_window() {
     // Add events within first window
     for i in 0..3 {
         let event = Event::new("SensorEvent")
-            .with_field("value", i as i64)
+            .with_field("value", i)
             .with_timestamp(base_time + Duration::seconds(i));
         engine.process(event).await.expect("Failed to process");
     }
@@ -123,7 +123,7 @@ async fn test_per_source_watermark_with_engine() {
     let base_time = Utc::now();
     for i in 0..5 {
         let event = Event::new("TestEvent")
-            .with_field("value", i as i64)
+            .with_field("value", i)
             .with_timestamp(base_time + Duration::seconds(i));
         engine.process(event).await.expect("Failed to process");
     }
